requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.4
pydantic-settings==2.9.1
numpy==2.2.5
sentence-transformers==4.1.0
faiss-cpu==1.11.0
google-genai==1.13.0
